requirements.txt
customtkinter==5.2.1
packaging 
pillow